backend/requirements.txt
fastapi
uvicorn
sqlmodel
psycopg2-binary
google-generativeai
python-dotenv
requests
pydantic-settings
